Used equality comparator for language option & Gathered the hashers in a class
