+++ lib/src/models/language_option.dart
@@ -27,4 +27,35 @@ class LanguageOption {
     required this.languageCode,
     this.iconPath,
   });
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+    if (other.runtimeType != runtimeType) return false;
+    return other is LanguageOption &&
+        other.value == value &&
+        other.languageCode == languageCode;
+  }
+
+  /// This hashCode part is inspired from Quiver package.
+  /// Quiver package link: https://pub.dev/packages/quiver
+  @override
+  int get hashCode => _finish(
+        _combine(
+          _combine(_combine(0, value.hashCode), languageCode.hashCode),
+          iconPath.hashCode,
+        ),
+      );
+
+  int _combine(int hash, int value) {
+    int newHash = 0x1fffffff & (hash + value);
+    newHash = 0x1fffffff & (newHash + ((0x0007ffff & newHash) << 10));
+    return newHash ^ (newHash >> 6);
+  }
+
+  int _finish(int hash) {
+    int newHash = 0x1fffffff & (hash + ((0x03ffffff & hash) << 3));
+    newHash = newHash ^ (newHash >> 11);
+    return 0x1fffffff & (newHash + ((0x00003fff & newHash) << 15));
+  }
 }
